+++ backend/src/main/java/com/deliveranything/domain/review/controller/ReviewController.java
@@ -46,7 +46,7 @@ public class ReviewController {
       @AuthenticationPrincipal SecurityUser user
   ) {
     if (!user.hasActiveProfile(ProfileType.CUSTOMER)) {
-      throw new CustomException(ErrorCode.ROLE_NOT_ALLOWED);
+      throw new CustomException(ErrorCode.PROFILE_NOT_ALLOWED);
     }
 
     ReviewCreateResponse response = reviewService.createReview(request, user.getId());
@@ -61,7 +61,7 @@ public class ReviewController {
       @PathVariable Long reviewId
   ) {
     if (!user.hasActiveProfile(ProfileType.CUSTOMER)) {
-      throw new CustomException(ErrorCode.ROLE_NOT_ALLOWED);
+      throw new CustomException(ErrorCode.PROFILE_NOT_ALLOWED);
     }
 
     reviewService.deleteReview(user.getId(), reviewId);
@@ -77,7 +77,7 @@ public class ReviewController {
       @RequestBody ReviewUpdateRequest request
   ) {
     if (!user.hasActiveProfile(ProfileType.CUSTOMER)) {
-      throw new CustomException(ErrorCode.ROLE_NOT_ALLOWED);
+      throw new CustomException(ErrorCode.PROFILE_NOT_ALLOWED);
     }
 
     ReviewResponse response = reviewService.updateReview(request, reviewId, user.getId());
