+++ backend/src/main/java/com/deliveranything/global/security/expression/ProfileSecurity.java
@@ -1,4 +1,4 @@
-package com.deliveranything.domain.store.store.security;
+package com.deliveranything.global.security.expression;
 
 import com.deliveranything.global.exception.CustomException;
 import com.deliveranything.global.exception.ErrorCode;
@@ -10,21 +10,21 @@ public class ProfileSecurity {
 
   public boolean isSeller(SecurityUser securityUser) {
     if (securityUser == null || !securityUser.isSellerActive()) {
-      throw new CustomException(ErrorCode.ROLE_NOT_ALLOWED);
+      throw new CustomException(ErrorCode.PROFILE_NOT_ALLOWED);
     }
     return true;
   }
 
   public boolean isCustomer(SecurityUser securityUser) {
     if (securityUser == null || !securityUser.isCustomerActive()) {
-      throw new CustomException(ErrorCode.ROLE_NOT_ALLOWED);
+      throw new CustomException(ErrorCode.PROFILE_NOT_ALLOWED);
     }
     return true;
   }
 
   public boolean isRider(SecurityUser securityUser) {
     if (securityUser == null || !securityUser.isRiderActive()) {
-      throw new CustomException(ErrorCode.ROLE_NOT_ALLOWED);
+      throw new CustomException(ErrorCode.PROFILE_NOT_ALLOWED);
     }
     return true;
   }
