+++ backend/src/main/resources/application.yml
@@ -91,7 +91,7 @@ oauth2:
 
 toss:
   payments:
-    secret-key: ${YOUR_TEST_SECRET_KEY}
+    secret-key: ${YOUR_TEST_SECRET_KEY:임시키}
     toss-url: https://api.tosspayments.com
 
 kakao:
@@ -100,11 +100,11 @@ kakao:
 
 custom:
   jwt:
-    secretKey: abcdefghijklmnopqrstuvwxyz0123456789abcdefghijklmnopqrstuvwxyz0123456789abcdefghijklmnopqrstuvwxyz0123456789
+    secretKey: ${JWT_SECRET_KEY:abcdefghijklmnopqrstuvwxyz0123456789abcdefghijklmnopqrstuvwxyz0123456789abcdefghijklmnopqrstuvwxyz0123456789}
   accessToken:
     expirationSeconds: "#{60*20}"
 
 gemini:
   api:
-    key: ${GEMINI_API_KEY}
+    key: ${GEMINI_API_KEY:임시키}
     url: https://generativelanguage.googleapis.com/v1beta/models/gemini-2.5-flash:generateContent
